Refactor config tools into separate file.
Also update/fix some of the tests in test_db.
